Manage /etc/elasticsearch/jvm.options for ES5
I'm using a combination of /etc/sysconfig/elasticsearch which is managed
by this formula and /etc/elasticsearch/jvm.options which is not.
I think it's best to give users the option to use all the config files,
so this simply populates jvm.options from Pillar. I've added the
defaults from jvm.options on a fresh installation to the example Pillar
too.

Closes issue #20.