Install specific version of ElasticSearch (#31)
ElasticSearch was mistakenly upgraded on one of the instances in my
cluster, which lead to problems. Luckily it only happened on QA.
This change allows us to pin specific versions of ES, avoiding that
problem.

The basic logic is....

If `version` pillar is set, strip out the `major_version` and set both
in the config map.  A specific package version is installed if
`version` is set... and we do lots of special things throughout the
formula if `major_version` is 5.

This should be compatible for those already using the formula even
though I'm deprecating the `major_version` pillar value.